Move the remanining two functions in the tools.py file to other source files
