+++ nw/common.py
@@ -311,3 +311,26 @@ def getGuiItem(theName):
         if qWidget.objectName() == theName:
             return qWidget
     return None
+
+def numberToRoman(numVal, isLower=False):
+    """Convert an integer to a roman number.
+    """
+    if not isinstance(numVal, int):
+        return "NAN"
+    if numVal < 1 or numVal > 4999:
+        return "OOR"
+
+    theValues = [
+        (1000, "M"), (900, "CM"), (500, "D"), (400, "CD"), (100, "C"), (90, "XC"),
+        (50, "L"), (40, "XL"), (10, "X"), (9, "IX"), (5, "V"), (4, "IV"), (1, "I"),
+    ]
+
+    romNum = ""
+    for theDiv, theSym in theValues:
+        n = numVal//theDiv
+        romNum += n*theSym
+        numVal -= n*theDiv
+        if numVal <= 0:
+            break
+
+    return romNum.lower() if isLower else romNum
