+++ nw/core/__init__.py
@@ -1,17 +1,15 @@
 # -*- coding: utf-8 -*-
 
 from nw.core.document import NWDoc
-from nw.core.index import NWIndex
+from nw.core.index import NWIndex, countWords
 from nw.core.project import NWProject
 from nw.core.spellcheck import NWSpellCheck, NWSpellEnchant, NWSpellSimple
 from nw.core.tohtml import ToHtml
 from nw.core.toodt import ToOdt
 from nw.core.tomd import ToMarkdown
-from nw.core.tools import countWords, numberToRoman
 
 __all__ = [
     "countWords",
-    "numberToRoman",
     "NWDoc",
     "NWIndex",
     "NWProject",
